test/e2e: Move the report expiration subtests to their own file
